requirements.txt
ultralytics
opencv-python-headless
numpy
streamlit
Pillow
